+++ influxdb_client/_sync/rest.py
@@ -101,9 +101,11 @@ class RESTClientObject(object):
                 cert_reqs=cert_reqs,
                 ca_certs=ca_certs,
                 cert_file=configuration.cert_file,
-                key_file=configuration.key_file,
+                key_file=configuration.cert_key_file,
+                key_password=configuration.cert_key_password,
                 proxy_url=configuration.proxy,
                 proxy_headers=configuration.proxy_headers,
+                ssl_context=configuration.ssl_context,
                 **addition_pool_args
             )
         else:
@@ -113,7 +115,9 @@ class RESTClientObject(object):
                 cert_reqs=cert_reqs,
                 ca_certs=ca_certs,
                 cert_file=configuration.cert_file,
-                key_file=configuration.key_file,
+                key_file=configuration.cert_key_file,
+                key_password=configuration.cert_key_password,
+                ssl_context=configuration.ssl_context,
                 **addition_pool_args
             )